requirements.txt
pandas>=1.5.0
numpy>=1.20.0
flask>=3.0.0
Flask-Babel>=4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0; sys_platform != 'win32'
jsonschema>=4.0.0